Address PR feedback, code cleanup
