+++ examples/test-site/src/components/FilterByGroup/ContextLogger.tsx
@@ -19,7 +19,7 @@ import {
 } from '@bodiless/fclasses';
 import {
   useFilterByGroupContext,
-  withRegisterSuggestions,
+  withSuggestionProps,
 } from '@bodiless/organisms';
 import { BodilessTag, TagType } from '@bodiless/core';
 
@@ -72,7 +72,7 @@ const ContextLoggerBase: FC<Props> = () => {
 };
 
 const ContextLogger = flow(
-  withRegisterSuggestions,
+  withSuggestionProps(),
 )(ContextLoggerBase);
 
 export default ContextLogger;
